requirements.txt
numpy~=1.24.4
pycuda~=2024.1
numba~=0.58.1
detectron2~=0.6
fvcore~=0.1.5.post20221221
pillow~=9.3.0
opencv-python~=4.5.5.64
pycocotools~=2.0.7
clip~=1.0
timm~=0.5.4
tabulate~=0.9.0
lvis~=0.5.3
nltk~=3.6.3
tqdm~=4.64.0
mss~=9.0.1
psutil~=5.9.8
torchvision~=0.17.2
pandas~=1.4.3
matplotlib~=3.5.2
setuptools~=68.2.2
scipy~=1.10.1
open3d~=0.15.2
iopath~=0.1.9
omegaconf~=2.2.2
yaml~=0.2.5
pyyaml~=5.4.1
dataclasses~=0.6
termcolor~=2.4.0
packaging~=24.0
cloudpickle~=3.0.0